Create Docx from output of auditfiles/requirements.txt
pandas>=1.5.0
python-docx>=0.8.11
openpyxl>=3.0.10